sources/Avalonia.GameStudio.Presentation/MarkupExtensions/LongExtension.cs
using System;

using Avalonia.GameStudio.Presentation.Internal;
using Avalonia.Markup.Xaml;

namespace Avalonia.GameStudio.Presentation.MarkupExtensions
{
    public sealed class LongExtension : MarkupExtension
    {
        public long Value { get; set; }

        public LongExtension(object value)
        {
            Value = Convert.ToInt64(value);
        }

        /// <inheritdoc/>
        public override object ProvideValue(IServiceProvider serviceProvider)
        {
            return Value;
        }
    }

    public sealed class MaxLongExtension : MarkupExtension
    {
        /// <inheritdoc/>
        public override object ProvideValue(IServiceProvider serviceProvider)
        {
            return Int64Boxes.MaxValueBox;
        }
    }

    public sealed class MinLongExtension : MarkupExtension
    {
        /// <inheritdoc/>
        public override object ProvideValue(IServiceProvider serviceProvider)
        {
            return Int64Boxes.MinValueBox;
        }
    }
}
